requirements.txt
diffusers==0.20.2
dill==0.3.7
einops==0.7.0
gradio==4.7.1
imageio==2.31.3
loguru==0.7.1
matplotlib==3.7.2
mmpose==1.2.0
neural_renderer_pytorch==1.1.3
numpy==1.23.1
omegaconf==2.3.0
open3d==0.17.0
opencv_python==4.8.0.76
Pillow==10.1.0
pyrender==0.1.45
pyrootutils==1.0.4
pytorch3d==0.7.4
pytorch_lightning==2.0.8
rich==13.7.0
scipy==1.11.4
setuptools==59.6.0
skimage==0.0
timm==0.9.7
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
transformers==4.33.1
trimesh==4.0.1
webdataset==0.2.57
yacs==0.1.8